chore(types): organize types
#53
