+++ src/components/DefaultComponents/DefaultLink/index.tsx
@@ -1,6 +1,12 @@
 import Link from "next/link";
 
-import { DefaultLinkProps } from "@/types/DefaultLinkProps";
+export interface DefaultLinkProps {
+  href: string;
+  titleAccessibility?: string;
+  hasPrefetch?: boolean;
+  cnLink?: string;
+  children: React.ReactNode;
+}
 
 export default function DefaultLink({
   href,
