--- src/types/DefaultLinkProps.ts
@@ -1,7 +1,0 @@
-export interface DefaultLinkProps {
-  href: string;
-  titleAccessibility?: string;
-  hasPrefetch?: boolean;
-  cnLink?: string;
-  children: React.ReactNode;
-}
