--- src/types/DefaultLogoProps.ts
@@ -1,3 +1,0 @@
-export interface DefaultLogoProps {
-  defaultLogoModifier?: string;
-}
