--- src/types/useWindowSizeProps.ts
@@ -1,4 +1,0 @@
-export interface useWindowSizeProps {
-  width: number;
-  height: number;
-}
